split to test diff files refactor
