+++ lib/main.dart
@@ -1,6 +1,6 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
+
+import 'dice.dart';
 
 void main() {
   return runApp(
@@ -19,43 +19,3 @@ void main() {
   );
 }
 
-class DicePage extends StatefulWidget {
-  @override
-  _DicePageState createState() => _DicePageState();
-}
-
-class _DicePageState extends State<DicePage> {
-  int leftDiceNumber  = 1;
-  int rightDiceNumber = 1;
-
-  void _rollDice() {
-    setState(() {
-      leftDiceNumber = Random().nextInt(6)+1;
-      rightDiceNumber = Random().nextInt(6)+1;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Row(
-        children: <Widget>[
-          Expanded(
-            flex: 1,
-            child: FlatButton(
-              onPressed: () => _rollDice(),
-              child: Image.asset('images/dice$leftDiceNumber.png'),
-            ),
-          ),
-          Expanded(
-            flex: 1,
-            child: FlatButton(
-              onPressed: () => _rollDice(),
-              child: Image.asset('images/dice$rightDiceNumber.png'),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
